wsj/s5/steps/nnet3/chain2/train.sh: added support to multilingual training

+++ egs/wsj/s5/steps/nnet3/chain2/train.sh
@@ -50,6 +50,7 @@ shuffle_buffer_size=1000  # This "buffer_size" variable controls randomization o
 
 
 l2_regularize=
+multilingual_eg=
 
 # End configuration section
 
@@ -115,6 +116,7 @@ echo "$0: will train for $num_epochs epochs = $num_iters iterations"
 # lrate and dropout_opt, among other variables.
 . <(head -n 1 $dir/schedule.txt)
 langs=$(awk '/^langs/ { $1=""; print; }' <$dir/init/info.txt)
+num_langs=$(echo $langs | wc -w)
 
 mkdir -p $dir/log
 
@@ -124,8 +126,14 @@ if [ $stage -le -1 ]; then
   # run.pl $dir/log/init_model_default.log \
   #     nnet3-am-copy  --learning-rate=$lrate $dropout_opt $dir/init/default.mdl $dir/0.mdl
   echo "$0: Copying transition model"
-  cp $dir/init/default.raw $dir/0.raw
-  cp $dir/init/default_trans.mdl $dir/0_trans.mdl
+  if [ $num_langs -eq 1 ]; then
+      cp $dir/init/default.raw $dir/0.raw
+  else
+      cp $dir/init/multi.raw $dir/0.raw
+  fi
+  if [ -f $dir/init/default_trans.mdl ]; then
+      cp $dir/init/default_trans.mdl $dir/0_trans.mdl 
+  fi
 fi
 
 
@@ -150,10 +158,12 @@ while [ $x -lt $num_iters ]; do
 
   next_x=$[$x+1]
   den_fst_dir=$egs_dir/misc
-  transform_dir=$dir/init
   model_out_prefix=$dir/${next_x}
   model_out=${model_out_prefix}.mdl
-
+  multilingual_eg_opts=
+  if $multilingual_eg; then
+       multilingual_eg_opts="--multilingual-eg=true"
+  fi
 
   # for the first 4 iterations, plus every $diagnostic_period iterations, launch
   # some diagnostic processes.  We don't do this on iteration 0, because
@@ -169,7 +179,7 @@ while [ $x -lt $num_iters ]; do
             $l2_regularize_opt \
             --print-interval=10  \
            "nnet3-am-init $dir/0_trans.mdl $dir/${x}.raw - | nnet3-am-copy --learning-rate=$lrate - - |" $den_fst_dir \
-           "ark:nnet3-chain-merge-egs --minibatch-size=$groups_per_minibatch scp:$egs_dir/${name}_subset.scp ark:-|" \
+           "ark:nnet3-chain-merge-egs $multilingual_eg_opts --minibatch-size=$groups_per_minibatch scp:$egs_dir/${name}_subset.scp ark:-|" \
            $dir/${next_x}_${name}.mdl || touch $dir/.error_diagnostic &
     done
   fi
@@ -195,7 +205,7 @@ while [ $x -lt $num_iters ]; do
              $l2_regularize_opt \
              --srand=$srand \
              "nnet3-am-init $dir/0_trans.mdl $dir/${x}.raw - | nnet3-am-copy --learning-rate=$lrate - - |" $den_fst_dir  \
-             "ark:nnet3-chain-copy-egs --frame-shift=$frame_shift scp:$egs_dir/train.$scp_index.scp ark:- | nnet3-chain-shuffle-egs --buffer-size=$shuffle_buffer_size --srand=$x ark:- ark:- | nnet3-chain-merge-egs --minibatch-size=$groups_per_minibatch ark:- ark:-|" \
+             "ark:nnet3-chain-copy-egs --frame-shift=$frame_shift scp:$egs_dir/train.$scp_index.scp ark:- | nnet3-chain-shuffle-egs --buffer-size=$shuffle_buffer_size --srand=$x ark:- ark:- | nnet3-chain-merge-egs $multilingual_eg_opts --minibatch-size=$groups_per_minibatch ark:- ark:-|" \
              ${model_out_prefix}.$j.raw || touch $dir/.error &
   done
   wait
@@ -239,18 +249,19 @@ if [ $stage -le $num_iters ]; then
   den_fst_dir=$egs_dir/misc
   input_models=$(for x in $(seq $combine_start_iter $num_iters); do echo $dir/${x}.raw; done)
   output_model_dir=$dir/final
-  transform_dir=$dir/init
 
    $cmd $gpu_cmd_opt $dir/log/combine.log \
       nnet3-chain-combine2 --use-gpu=$use_gpu \
         --leaky-hmm-coefficient=$leaky_hmm_coefficient \
         --print-interval=10  \
         $den_fst_dir $input_models \
-        "ark:nnet3-chain-merge-egs  scp:$egs_dir/train_subset.scp ark:-|" \
+        "ark:nnet3-chain-merge-egs $multilingual_eg_opts  scp:$egs_dir/train_subset.scp ark:-|" \
         $dir/final.raw || exit 1;
-   nnet3-copy  --edits="rename-node old-name=output new-name=output-dummy; rename-node old-name=output-default new-name=output" \
-      $dir/final.raw - | \
-      nnet3-am-init $dir/0_trans.mdl - $dir/final.mdl
+   if [ "$langs" == "default" ]; then
+       nnet3-copy  --edits="rename-node old-name=output new-name=output-dummy; rename-node old-name=output-default new-name=output" \
+          $dir/final.raw - | \
+          nnet3-am-init $dir/0_trans.mdl - $dir/final.mdl
+   fi
 
 fi
